Fixes in unit tests.
